+++ Cargo.toml
@@ -25,6 +25,7 @@ id_tree = "1.8.0"
 lazy_static = "1.4.0"
 once_cell = "1.19.0"
 parking_lot = "0.12.1"
+pretty-bytes = "0.2.2"
 prettyplease = "0.2.16"
 proc-macro2 = "1.0.76"
 quote = "1.0.35"
