+++ gdsfx-app/Cargo.toml
@@ -17,6 +17,7 @@ gdsfx-build = { path = "../gdsfx-build" }
 gdsfx-data = { path = "../gdsfx-data" }
 gdsfx-library = { path = "../gdsfx-library" }
 once_cell = { workspace = true }
+pretty-bytes = { workspace = true }
 rust-i18n = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
