+++ gdsfx-app/src/tabs/tools.rs
@@ -1,0 +1,7 @@
+use eframe::egui::{Ui, Context};
+
+use crate::GdSfx;
+
+pub fn render(ui: &mut Ui, gdsfx: &mut GdSfx, ctx: &Context) {
+    todo!()
+}
